refactor: Images, Posts, Users Entity 수정
