+++ src/main/java/com/example/ReviewZIP/domain/postHashtag/PostHashtags.java
@@ -1,7 +1,7 @@
 package com.example.ReviewZIP.domain.postHashtag;
 
 import com.example.ReviewZIP.domain.hashtag.Hashtags;
-import com.example.ReviewZIP.post.Posts;
+import com.example.ReviewZIP.domain.post.Posts;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
