+++ src/main/java/com/example/ReviewZIP/domain/scrab/Scrabs.java
@@ -2,7 +2,7 @@ package com.example.ReviewZIP.domain.scrab;
 
 import com.example.ReviewZIP.domain.user.Users;
 import com.example.ReviewZIP.global.entity.BaseEntity;
-import com.example.ReviewZIP.post.Posts;
+import com.example.ReviewZIP.domain.post.Posts;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
