+++ src/main/java/com/example/ReviewZIP/domain/user/Users.java
@@ -1,6 +1,7 @@
 package com.example.ReviewZIP.domain.user;
 
 import com.example.ReviewZIP.domain.follow.Follows;
+import com.example.ReviewZIP.domain.post.Posts;
 import com.example.ReviewZIP.domain.scrab.Scrabs;
 import com.example.ReviewZIP.global.entity.BaseEntity;
 import jakarta.persistence.*;
@@ -40,6 +41,9 @@ public class Users extends BaseEntity {
     @Column(nullable = false)
     private String nickname;
 
+    @Column(nullable = false)
+    private String profileUrl;
+
     @Enumerated(EnumType.STRING)
     private Status status;
 
@@ -51,4 +55,7 @@ public class Users extends BaseEntity {
 
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
     private List<Scrabs> scrabList = new ArrayList<>();
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+    private List<Posts> postList = new ArrayList<>();
 }
